Sushruta_Model/requirements.txt
# data process
pandas
numpy
albumentations

# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.5.10
torchmetrics>=0.7.0
# timm

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
hydra-optuna-sweeper>=1.1.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# tensorboard

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- others --------- #
gdown           # for download data from google drive
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
pudb            # debugger

# cholect45 evaluation
ivtmetrics